requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-MQTT==1.2.1
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paho-mqtt==1.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.18.1